third_party/WebKit/LayoutTests/fast/table/table-rowspan-height-less-than-one-percent-expected.txt

Test for chromium bug : 408066. ASSERTION FAILED: !extraRowSpanningHeight in blink::RenderTableSection::distributeRowSpanHeightToRows.
FAIL:
Expected 22 for height, but got 20. 

<th id="perc-height" data-expected-height="22">
            </th>
r0c0
r2c0
FAIL:
Expected 24 for height, but got 22. 

<th id="perc-height" data-expected-height="24">
            </th>
r0c0	r0c1
r2c0
